requirements.txt
gradio==5.38.0
#llava.egg==info
matplotlib==3.10.3
mmcls==0.25.0
mmcv==2.0.0
mmcv_lite==2.2.0
mmdeploy==1.3.1
mmdet==3.0.0
mmengine==0.10.7
mmpose==1.3.2
mmrazor==1.0.0
mmrotate==0.3.4
nltk==3.9.1
numpy==2.3.1
onnx==1.18.0
onnxruntime==1.22.0
onnxsim==0.4.36
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.1
parameterized==0.9.0
Pillow==11.3.0
prettytable==3.16.0
pycocoevalcap==1.2
pycocotools==2.0.10
pytest==8.4.1
pytorch_sphinx_theme==0.0.19
PyYAML==6.0.2
rich==14.0.0
rouge_score==0.1.2
sacrebleu==2.5.1
sahi==0.11.31
scipy==1.16.0
seaborn==0.13.2
setuptools==60.2.0
Shapely==2.1.1
supervision==0.26.0
tensorflow==2.19.0
tensorrt==10.12.0.36
timm==1.0.17
torch==2.7.1
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.22.1
torchviz==0.0.3
tqdm==4.65.2
